+++ lib/transformNestedSelection.d.ts
@@ -1,7 +1,8 @@
-import { GraphQLResolveInfo } from 'graphql';
+import { DocumentNode, GraphQLResolveInfo } from 'graphql';
 import { Variables } from './types';
-declare const transformNestedSelection: (info: GraphQLResolveInfo, prefix: string) => {
-    document: any;
+declare const transformNestedSelection: (info: GraphQLResolveInfo, prefix: string, wrapper?: DocumentNode | undefined) => {
+    document: DocumentNode;
     variables: Variables;
+    wrappedPath: string[] | undefined;
 };
 export default transformNestedSelection;