+++ lib/types.d.ts
@@ -51,4 +51,11 @@ export interface DataLoaderKey<TContext> {
     };
 }
 export declare type EndpointDataLoader<TContext> = DataLoader<DataLoaderKey<TContext>, EndpointExecutionResult>;
+export interface NestedSource<TContext> {
+    context: TContext;
+    wrapper?: DocumentNode;
+    errors?: BaseGraphQLError[] | null;
+    errorPromise?: Promise<BaseGraphQLError[] | null | undefined>;
+    resolveErrors?: (errors: GraphQLEndpointError[] | null | undefined) => void;
+}
 export {};