+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # nest-graphql-endpoint
 
+## 0.1.0
+
+### Added
+- field resolver functionality
+- githubRequest on nestGitHubEndpoint
 ## 0.0.2
 
 ### Fixed